# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0",
        "description": "Documentation of API endpoints for the application"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur Local"
        }
    ],"components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        
        }
    },
    "security": [
        {
            "BearerAuth": []
    
            }
        ],
    "paths": {
...
        "/api/v1/apprenant": {
            "post": {
                "summary": "Create a new apprenant",
                "description": "This endpoint is used to create a new apprenant and associate them with a referentiel and promotion.",
                "operationId": "createApprenant",
                "tags": [
                    "Apprenants"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {

                                    "nom": {
                                        "type": "string"
                                    }, 
                                    "adresse": {
                                        "type": "string"
                                    },

                                    "prenom": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "referentiel_id": {
                                        "type": "string"
                                    },
                                    "tuteur_nom": {
                                        "type": "string"
                                    },
                                    "tuteur_prenom": {
                                        "type": "string"
                                    },
                                    "tuteur_contact": {
                                        "type": "string"
                                    },
                                    "cni": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "diplome": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "extrait_naissance": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "visite_medicale": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "casier_judiciaire": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "statut": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "nom": {
                                            "type": "string"
                                        },
                                        "prenom": {
                                            "type": "string"
                                        },
                                        "contact": {
                                            "type": "string"
                                        },
                                        "photo": {
                                            "type": "string"
                                        },
                                        "referentiel_id": {
                                            "type": "string"
                                        },
                                        "tuteur_nom": {
                                            "type": "string"
                                        },
                                        "tuteur_prenom": {
                                            "type": "string"
                                        },
                                        "tuteur_contact": {
                                            "type": "string"
                                        },
                                        "cni": {
                                            "type": "string"
                                        },
                                        "diplome": {
                                            "type": "string"
                                        },
                                        "extrait_naissance": {
                                            "type": "string"
                                        },
                                        "visite_medicale": {
                                            "type": "string"
                                        },
                                        "casier_judiciaire": {
                                            "type": "string"
                                        },
                                        "statut": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },"/api/v1/apprenants/{id}": {
...
"/api/v1/apprenants": {
    "get": {
        "summary": "List all apprenants",
        "description": "This endpoint returns a list of all apprenants in the system.",
        "operationId": "listApprenants",
        "tags": [
            "Apprenants"
        ],
        "responses": {
            "200": {
                "description": "Successful response",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": { "type": "string" },
                                    "nom": { "type": "string" },
                                    "prenom": { "type": "string" },
                                    "telephone": { "type": "string" },
                                    "referentiel_id": { "type": "string" },
                                    "photo": { "type": "string" },
                                    "statut": { "type": "string" }
                                }
                            }
                        }
                    }
                }
            },
            "500": {
                "description": "Internal Server Error"
            }
        }
    }
},
"/api/v1/apprenants/import": {
    "post": {
        "summary": "Import multiple apprenants",
        "description": "This endpoint allows importing multiple apprenants via file upload.",
        "operationId": "importApprenants",
        "tags": [
            "Apprenants"
        ],
        "requestBody": {
            "required": true,
            "content": {
                "multipart/form-data": {
                    "schema": {
                        "type": "object",
                        "properties": {
                            "file": {
                                "type": "string",
                                "format": "binary",
                                "description": "The file containing apprenant data to import."
                            }
                        }
                    }
                }
            }
        },
        "responses": {
            "200": {
                "description": "Successful import",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": { "type": "string" },
                                "importedCount": { "type": "integer" }
                            }
                        }
                    }
                }
            },
            "400": {
                "description": "Bad Request"
            },
            "500": {
                "description": "Internal Server Error"
            }
        }
    }
}
  }
}